refactor: split RateLimitFetcher into focused extensions (743 → 494 lines)
`AIBattery/Services/RateLimitFetcher.swift` was at 743 lines. Three
self-contained method groups peeled off into same-type extensions:

- RateLimitFetcher+UsageEndpoint.swift (100 lines) — the dedicated
  `/api/oauth/usage` primary path. `interpretUsageEndpoint` (the
  nonisolated static pure interpreter pinned by 8 of the contract tests
  added in v2.3.2) and the async `fetchUsageEndpoint` wrapper that
  delegates to it.

- RateLimitFetcher+ClientData.swift (126 lines) — the Claude Code
  `/api/oauth/claude_cli/client_data` fallback path.
  `interpretClaudeCodeClientData` (pure, the other half of the v2.3.2
  contract suite), the async `fetchClaudeCodeClientData` wrapper, and
  `containsStandardRateLimitHeaders`. The latter is also called from
  the Messages-API path in the main file, so it's dropped from
  `private static` to `static`.

- RateLimitFetcher+Persistence.swift (65 lines) — UserDefaults-backed
  per-account cache survival: `PersistedRateLimits` (`fileprivate`,
  visibility scoped to the extension file) plus
  `persistRateLimits` / `restorePersistedRateLimits`.

What stayed in the main file:

- The orchestrating `fetch(...)` entry point, the public `cachedOrEmpty`
  and `setCachedResult` API, observed/working model bookkeeping
  (`restoreWorkingModels`, `setObservedModels`, `saveWorkingModel`),
  the pure helpers `parseRetryAfter` / `quotaThrottleLikely` /
  `quotaExhaustionThreshold`, and the Messages-API path itself
  (`buildHeaderResult` + `tryFetch`) — 230 lines of the original 743,
  the densest concentration of branches, and still the right place to
  reason about the multi-stage fallback ladder.

Cross-file visibility widened to `internal` (module-private), each
annotated inline:

- `cachedResults`, `persistKeyPrefix` (used by persistence)
- `usageURL`, `messagesURL`, `clientDataURL`, `userAgent` (used by both
  endpoint extensions and the Messages-API path)
- `saveWorkingModel(_:accountId:)` (called from client_data on success)

Nothing outside the module touches any of these — they remain
package-private.

No production behavior change. Test suite: 1002/1002 passing (Test run
with 1002 tests in 69 suites passed after 2.452 seconds).

After:
- RateLimitFetcher.swift:                494 lines (was 743)
- RateLimitFetcher+ClientData.swift:     126 lines
- RateLimitFetcher+UsageEndpoint.swift:  100 lines
- RateLimitFetcher+Persistence.swift:     65 lines
- (combined): 785 lines (was 743, +42 from per-file imports/comments)
